Remap PUA characters to original parsed data
When parsing data and writing to XML, we map characters that are illegal
in XML to PUA. When reading this XML back in for unparse, we need to map
PUA characters back to the original values. This adds that function, and
refactors the remap function to reuse the string traversal code.

This also adds conversion for INF, -INF, and NaN for double and float
types to the PrimTypes that was added in InfosetImpl in commit
7bf1c52eadffbf6bbb6b4cfc003ce1aa3340bec5.

DFDL-1369