Support end events for determine unparse choice branches
In some cases, it is an end event that determines which choice branch to
unparse. For example:

  <element name="a">
    <choice>
      <element name="a">
      <sequence />
    </choice>
  </element>

In this case, the infoset could look like:

  <a><a></a></a>

and seeing the start event for "a" would mean we choose the "a" branch of
the choice. However, the infoset could also look like:

  <a></a>

and seeing the end event for "a" (note a different "a", but the same
qname) would mean we choose the empty sequence branch of the choice.

Currently we only match against start events, and we would get an error
when trying to find identifying elements for this empty sequence.

This change makes it so that we also find identifying elements
for a choice that are end events, and modifies the choice unparser to
match both start and end events.

DFDL-1313