Ensure ElementRef always uses the ElementRuntimeData from the referenced element
The runtimeData of an element and an element reference are virtually the
same, except for things like schema location. However, these runtime
data are used to assert that the correct events are received during
unparse. If an event uses an element and the parser uses a reference, or
vice cersa, the assertions fail because the expected erd doesn't exactly
match the actual. So modify ElementRef to always get the runtime data
from the referenced element to ensure these assertions pass.

DFDL-1398