Verified that the @AfterClass, even if put in the object, not the class, is run after all the tests are complete, and is not just run when the jvm shuts down, but at completion of those tests.
Changed Runner to be thread-safe. Can now run tests in parallel (should
be able to)

Changed TDMLRunner to use a thread-safe collection for the compiled
schema cache.

Generic SchemaCache, unit tests of it. Recompiles if file has changed.

Synchronized block around compiler, so we're only running one compile of
any schema at a time. I don't know that this is necessary, but it is the
conservative thing to do, as we've not tried to make compilation thread
safe.

DFDL-1300