Removed dependency on the ICU CharsetICU and related classes.
These were painfully requiring us to have a manually maintained lib
directory with two manually maintained jar files.

And anything that used Daffodil, also needed these.

So this dependency is now gone. We're still using ICU, just not the
parts that were in these unmanaged jars.

Turns out the only thing we had that was depending on these unmanaged
jars was some unit tests.

Interesting that two tests in daffodil-test changed behavior as a
result. They now insert two Unicode replacement characters instead of
one.

This happened due to a change in UTF-8 decoder behavior. What it does is
different when it encounters broken data that starts with F0 - the utf-8
indicator byte for a 4-byte codepoint. It doesn't get an error unless
there is room in the output char buffer for 2 16-bit codepoints. 

This required some adaptation in the fillCharBuffer call of the
dataInputStream layer.

DFDL-1020