Remove scalatest as a dependency
Having two test infrastructures (junit and scalatest) caused confusion
in some cases. For example, if you ever tried to run sbt testOnly and
provide arguments, the arguments would be passed to both test
infrastructures. However, the arguments are specific to the
infrastructures, which the other didn't know about, which caused a
failure. So remove the scalatest dependency and convert those tests
(only passera) to junit.

DFDL-1404