Replace NS thread cache with UniquenessCache from scala
The existing cache using a mutable map was not thread safe. Replace that
with the UniquenessCache from scala which contains all the logic to be
thread safe and simplifies the NS code a bit. This also changes the
cache to key on a URI rather than a string representation of a URI.

DFDL-1408