Added MStack class which is @specialized for all our stacks of Long and Boolean - without this, every push/pop of a Boolean or a Long was Boxing and unboxing, that is allocating.
Also MStack adds a mark/reset capability for resetting a stack, and
removes most of the operations on stacks that were O(n).

Eliminated all @deprecated methods, and unnecessary copying of PState,
UState, and the DataInputStream and DataOutputStream. The mark objects
used to implement mark/reset are all managed in pools now, not allocated
on the heap.

Put back build.sbt setting that converts warnings to errors.

Abstracted out common Pool utility class

Abstracted out common ThreadCheckMixin utility

Eliminated some use of ThreadLocal in PState and BBDataInputStream.
All the state that was in those ThreadLocal objects is now part of PState or
BBDataInputStream directly. The ThreadLocal is just there to cache large
buffers per thread so we don't allocate these large buffers over and over when
testing on 20K files for example.

Created Debugger-specific state-capture object that is used to copy bits
of state explicitly needed by the debugger. This eliminates the need for
the PState.Mark object to carry things that aren't needed for mark/reset
but the debugger was accessing.

Added markPos and resetPos which are mark/reset for just the position.

Modified DFA and specified length and assert pattern parsers to use
this. Should improve performance/decrease allocation overhead for
mark/reset.

DFDL-1415, DFDL-1417