Optimize enumeration value lookup
In some cases we look up the value of a enum by string at runtime, like
with PCAP when looking up the result of the byteOrder expression and
converting "bigEndian" or "littleEndian" to the byte order enum value.
With PCAP, we this many times since we can't yet statically determine if
the when the value of the expression might change and so we reevaluate
the expression many times, and thus reevaluate the byte order.

The string to enum value algorithm is innefficient for what it does. It
first converts to lowercase (a char array allocation), hashes the
string, and looks up in a hash table. Considering that that the vast
majority of enum's have less than 3-4 values, a map look up is overkill.

So replace the Map with an ArrayBuffer and just linearly look through
the array using a while loop. This removes the hash calculation, as
well as removes the Option allocation returned from the map lookup.
Additionally, use equalsIgnoreCase to avoid a char array allocation.

In testing of PCAP, this resulted in about a 12% improvement in
performance.

DFDL-1421