Introduced Cursor, XMLEventCursor
This is about generally getting excess allocation and copying out of the
low-level of the unparser.

Removed UnparserAugmentingInfosetSource.

It just cannot work. It walks the schema, sure, but doesn't evaluate
variables, nor put variables into scope, etc.

Also, since expressions can use dfdl:valueLength and dfdl:contentLength
it isn't possible to augment with calculated values until actually in
the serializer part.

You could perhaps do defaulting, but a whole pass just for that is
silly.

Removed XMLInfosetEventIterator.

This was doing a full pass over the incoming infoset XML events to
synthesize an intermediate thing called an XMLInfosetEvent which is not
quite an InfosetEvent.

Got rid of XMLInfosetEvent, and XMLInfosetEventIterator (used to be
called XMLInfosetIterator)

Re-wrote so that InfosetSource works directly off of XMLEventCursor and
is itself a Cursor, not an Iterator or Stream.

This should be much more efficient.

Also applied as many notions of performance programming (that have been
learned trying to improve the parser performance.)

Added unit tests for InfosetCursorFromTree.

Cleaned up deprecated CursorFromIterator hack.

UState is no longer an iterator.

Got rid of spurious println in many tests.