Improvements per review.
Renaming to use word "Cursor and Accessor" appropriately.

Moved IteratorFromCursor into daffodil-lib-unittest,
rather than have it be deprecated.
Lots of tests have to be edited otherwise.
Similarly the Start and End objects (which have unapply for pattern
matching - that's ok in tests)

Added typed eq/ne named _eq_ and _ne_. These prevent a whole bunch of
errors in the runtime where one accidently compares by eq (for
efficiency), say an infoset node, to an ERD by accident.

Removed need for DPathCompileInfo in InfosetImpl.scala - cleans up this
layering between The DPathCompileInfo and the ElementRuntimeData
objects. See the comments on the DPathCompileInfo trait/class.

Coroutine improvements. Doesn't catch Exception or RuntimeException.

Removed these dangerous catches. Users of InvertControl must put in
their own catches, and propagate them to the consuming side for any
application-specific exceptions. Other exceptions cause the producer to
exit, but the consumer side is notified with a generic wrapper-like
exception and it will abort when so notified.