Revised VariableMap to be more abstract.
Uses VariableRuntimeData now as the primary way to index into a map.

For debug purposes, you can still use a globalQName to index in. Since
that's what unit tests have around more easily.

This implementation could now be changed into one that is a mutable part
of pstate/ustate, and we could eliminate the hash-table/map and all the
allocation and cost and replace it with an array where a variable VRD
provides a slot number for use at runtime.

Note that backtracking must unwind variable state changes.