Reduce allocations of complex state
When saving DIComplex state, rather than allocating a new array the same
size as the number of slots and setting flags in that array, we can
instead just keep track of the last slot that was updated (and if it was
an array, how big it was). This is all the information necessary to
backtrack properly. Note that this will not work with unordered
sequences. For that, we will need a new DIComplex type that does record
all the necessary state, but we can implement that when unordered
sequences are reenabled.

On various tests, these changes saw about a 10% increase in performance

DFDL-1464