Remove mandatory marking in StatementElementCombinator
StatementElementCombinator currently marks at the beginning of the parse
function, regardless of if it is actually necessary or not. And it is
only actually necessary if there are pattern discriminators or
assertions. If those do not exists, then we never need to save the state
in this parser. Refactor the code so that marks are avoided when
possible.

This also discovered a bug in which discarding a PState.Makr did not
properly discard the associated data input stream mark.

Additionally, change some Option[Parser]s and Array[Parser]s with
size 1 to Maybe[Parser]s.

Initial testing saw a 20-25% increase in performance.

DFDL-1465