Rename tutorial tdml to have ".xml" extension.
The tutorial files are going to want to be served from web servers we
don't control and can't customize. Mostly these set the MIME type of the
data based on the file extension, and they won't know what ".tdml" is so
it will come through as type text, and won't get XSL transformed.

If we use ".xml" extension on the end, then such servers will get the
MIME type right and all will work.

So tutorial files naming convention is topic.tutorial.tdml.xml