OutputNewLine converted to "Ev" idiom.
Deprecation warnings added to values associated with the old way of
doing things. This results in 66 warnings that I'm not planning to fix
in this commit, hence, I've changed the build.sbt to tolerate warnings
temporarily.

Charsets are standardized by cooker to uppercase trimmed.

And other changes per code review. With the exception of the conversion
of runtimeDependencies manually maintained to an automated
reflection-based approach. This will be investigated separately.

DFDL-1271