Establish the asymetry between DataInputStream and DataOutputStream
The primary contribution of this commit is to establish the asymetry
between DataInputStream and DataOutputStream with respect to the
positions and limits in the bit stream.

For the DataInputStream, we have this concept of the current bitPos0b,
and optionally there may be abound called bitLimit0b. There are 1b
variants of these.

For parsing, these are always absolute values, that is they contain bit
position relative the ultimate start of the input stream where parsing
began.

For DataOutputStream, we have slighly different concepts.

There are absolute and relative variants. The absolute bitPosOb or
absBitPos0b is symmetric to the parser's bitPos0b. It's the position
relative to the ultimate start of the output stream.

However, we often do not know this value. So the UState and
DataOutputStream have a maybeAbsBitPos0b which can be MaybeULong.Nope if
the value isn't known.

In addition we have the relative or relBitPos0b. This is relative to the
start of whatever buffer we are doing unparsing into.

When unparsing, we often have to unparse into a buffer where the
ultimate actual absolute position isn't yet known, but we have to do the
unparsing anyway, for example so that we can measure exactly how long
something is.

Conversely, sometimes we simply must have the absolute output bit
position, for example, when computing the number of bits to insert to
achieve the required alignment.

Hence we have relBitPos0b - always known and is a value >= 0, and we
have maybeAbsBitPos0b which is a MaybeULong. If known it is >=0.

Corresponding to bit position we have bit limit, which is measured in
the same 0b or 1b units, but is *always* a maybe type, because even in
the case where we know the absolute position, we still may or may not
have any limit in place. Hence the UState and DataOutputStream have a

maybeRelBitLimit0b
and
maybeAbsBitLimit0b.

One invariant is this: when the absolute bit pos is known, then it is
the same as the relative bit pos. Similarly when the absolute bit limit
is known, then the relative bit limit is known and is equal.

All deprecation warnings fixed.

No warnings now. Turned back on -Xfatal-warnings in scala compile
options.

There is still CachedDynamic in here, which is now using Evaluatable, so
we have the caches on top of the caches still.

Added missing maybeAbsBitPos0b and 1b methods.