Fixes issue where unparser tests that expect SDEs were failing.
Runtime code on unparser side was missing
some case statements to handle SDEs.

Essentially, prior to the fix for DFDL-261
these SDEs were not being caught and were
simply bubbling up.  Unfortunately, now that
we have added the missing cases, we can no
longer use the 'interecept' code to catch
the exceptions.  Exceptions now end up
in a diagnostics list, which then ends up
being rethrown as a new Exception.  This
would cause the intercept code to fail
because it expected SuspendedExpressionsDeadlockException
but got java.lang.Exception.

For the failing test, testoutputValueCalcDeadlock,
we use a try-catch instead of the intercept command.
This corrects the issue.

Moved ProcessingError out of ParseErrors file into
its own file so that UnparseErrors can refer to it.

DFDL-1489