The test for DFDL-930 passes already. This contains the fix for DFDL-929.
The issue came from two problems:
1. That we were returning within the
parameter of a function (ExecutionMode).  This caused
a NonLocalReturnControl to be thrown.

Unfortunately the NonLocalReturnControl error contains
no stack trace.  So there was no obvious way to trace
its origin.

2. The expected infoset element name was incorrect in
the TDML document.  This was made apparent once we
were able to force the stack trace using the Scala
system property: scala.control.noTraceSuppression=false.

The lesson learned here is that we should avoid use of
the return statement as it is:
1. Hard to find the source.
2. Hides the real errors.

DFDL-930, DFDL-929