Ensure all marks are discarded, even if exceptions are thrown
If an unexpected exception was thrown sometime during parse, it could
result in not resetting/discarding marks. The exception would eventually
bubble up to the top, but before it got to the very top, we would check
to make sure everything in the mark pool was cleared. Because of the
exception, it was possible the mark pool wasn't cleared, so we would
immediately abort with an invariant, thus masking the original exception
that caused the problem. This makes it very difficult to debug
unexpected exceptions.

This changes Daffodil so that at the very top in doParse, we only
validate the final state if no unexpected exceptions are thrown. If we
see any unexpected exceptions, we immediately throw them, and skip
validation.

Additionally, remove a catch of generic Exception in the
OccursCountKindExpression parser. We should never be catching an generic
exception without rethrowing.

Also, avoid complex processing in a finally block in parse(). If
something fails in the finally blow and throws an exception, it will
mask the original exception (if one was thrown in the catch).

DFDL-1483