Progress toward dfdl:outputValueCalc and dfdl:contentLength function.
Removed DISequence and DIChoice which are no longer needed.
Preserved code and test code for possible use in the future, for
experimental feature when dfdl:outputValueCalc or similar feature is
extended to allow computing complexType elements.

dfdl:valueLength - stubbed.

DataOutputStream - these now preserve the absolute bit pos information
(for use in alignment regions) as buffers are spun off. The length of an
OVC element is added in so that we have the correct absolute bit pos at
the start of a buffer.

Infoset modified to have places to store starting and ending bitPos
for every node. This can be used to compute contentLength in bits/bytes.

Infoset throws specific Infoset exceptions if a
request is made to get the contentLength or valueLength, and it has not
yet been computed. This is the mechanism by which OVC expressions can
block their evaluation to wait for this information to be created.

The DPath expression operations that access the infoset do so by way of
methods on the DState object. When in Blocking mode (for forward referencing
expressions) these implement a retry-loop if the infoset throws exceptions.
Coroutines are used to allow unparsing to continue while waiting for this
to be retried.

Modified Element parsers and unparsers to update the start and end bitPos
of infoset elements. All element combinators changed so that they get a
before, element, and after gram. So that the processing of the content
(without any framing) is easily isolated. That's where the contentLength
-in bits, is captured.

(Note: bug - there have to be separate contentLength slots for parsing
and unparsing. Otherwise when parsing something can have zero
contentLength, a default value gets created, and on unparsing it has a
non-zero content length. If we round-trip and unparse from a
parser-created infoset, then an OVC element that reaches forward to get
the contentLength will get the answer left there by the parser, rather
than the expression blocking for the unparser to fill in the
contentLength. We could introduce separate content length for use by
parser and unparser, or we can clear the content-length as infoset
events are processed. )

Split OVC grammar primitive into a static and runtime flavor.

(TODO: create a lengthInBits EV that depends (optionally) on the
LengthEv) Then this could be collapsed back into just one flavor.

SuspendableExpessions use Coroutines.

Alignment is not possible after a variable-length OVC element
(for now), because alignment suspensions are not yet implemented.

Updated some tests so they will pass stricter checking.

Added unit tests for some middle end attributes that are critical to
getting nextElementResolver right.

Several tests of forward-expressions calling dfdl:contentLength are in
daffodil-test section 23 dfdl-expressions, contentLength.tdml.

Ignore hiddenGroupRefs when calculating possible next/firstChild elements

Properly convert text date/times from XMLCursor to DFDLDateTimes

Disable round tripping text bigInteger. BigInteger doesn't work properly
on unparse. When the ICU number formatter formats a bigInt, it looses a
lot of precision. For example, this

     BigInt(686479766013060971498190079908139)

Formats to the string:

     "686,479,766,013,061,000,000,000,000,000,000"

Updates to PCAP to use outputValueCalc for ICMP packets, including
adding support for EchoRequest and EchoReply in ICMP PCAP messages

Make OVC elements optional in the infoset when unparsing.
- Modify the next element resolvers to treat OVC elements as optional
- Modify the OVC statement element combinator to optionally consume OVC
  element events if they exist

Fix unparsing recurrence indicators expressions for mil-std-2045

    Recurrence indicator expressions in mil-std-2045 look like this:

      dfdl:outputValueCalc="{ if (dfdl:occursIndex() lt fn:count(..)) then 1 else 0 }

    Two issues were fixed to support this:

    - dfdl:occursIndex() was always 1 due to the way we cloned UState for
      OVC. We also clone the dState, but upon expression evaluation, some
      dState values are overwritten with values in the UState. This is the
      case with arrayPos. So when cloning the UState, clone the stack that
      is used to set arrayPos in the dState upon expression evaluation.

    - The parent axis (..) only moves up to the parent, which is always
      going to be a complex type. However, used in the specific context of
      fn:count, we really want the parent axis to reference the actual
      array so we can get it's length. To support this, modify DPath so
      that if .. is the last step element, and that element is an array, and
      the target type of that path is an array, then use a different dpath
      recipe to access the array rather than the complex type. This also
      requires adding a new field to DIElements to store the array if they
      are in one.