Set the current infoset node in the StatementElementUnparserNoRep
Even though these elements are inputValueCalc, we still need to set the
infoset node. This is for a couple of reasons:

1) The StatementElementUnparserBase asserts that it has an infoset. As a
   contrived example, imagine a schema in which the root is an
   inputValueCalc element. There will only ever be one infoset node, so
   the IVC unparseBegin/End must set it so this invariant passes. Note
   that although this example is contrived, it is used heavily in TDML
   tests, and is technically valid DFDL, so it should work.

2) Imagine an IVC element that sets a variable, and that setVariable
   expression references the value of the IVC element. In this case, the
   current info node does need to be set to the IVC infoset element so
   that the setVariable expression can access it with '.' or
   'self::foo'.

So, this modifies StatementElementUnparserNoRep to use the same
unparseBegin/End functions as StatementElementUnparser (i.e. do not
override them) so that the current infoset node is set to the IVC when
unparsing. Now the only difference between StatementElementUnparser and
NoRep, is move() override to avoid separators from being laid down.

This change enables about 150 extra tests to round trip.

DFDL-1543