Add support for daf:parseUnparsePolicy
This daffodil extension property determines whether or not we should
build a parser, an unparser, or both. Valid values for this property are
"parseOnly", "unparseOnly", or "both". This is used to determine if
Daffodil should build a parser, an unparser, or both. Additionally,
disables SDE checking that are specific parse or unparse if either are
disabled. This property can be used like so:

First define the daf extension namespace:

  xmlns:daf="urn:ogf:dfdl:2013:imp:opensource.ncsa.illinois.edu:2012:ext"

Then set the property simliar to how normal dfdl properties are set, but
include the daf prefix:

  <dfdl:format daf:parseUnparsePolicy="parseOnly" ... />

  <xs:element daf:parseUnparsePolicy="parseOnly" ... />

  <xs:annotation>
    <xs:appinfo source="http://www.ogf.org/dfdl/">
      <dfdl:element>
        <daf:property name="parseUnparsePolicy">parseOnly</daf:property>
      </dfdl:element>
    </xs:appinfo>
  </xs:annotation>

Note that to use the element form, the property element is in the daf
namespace.

To determine what should be be built and checked, we look at the
parseUnparsePolicy of the root element. All children elements must have
a compatible parseUnparsePolicy or it is a SDE. For a child element to
be compatible, it must either have the same parseUnparsePolicy as the
parent, or it must have a "both" parseUnparsePolicy.

This also adds a tunable (parseUnparsePolicy) that can have values of
'parseOnly', 'unparseOnly', 'both', or 'schema'. If the value is
'schema' (the default value), then the policy is determined by looking
at the root element. Otherwise, the root element is ignored and the
value of the tunable is used. All children, including the root, must be
compatible with the tunable policy if it is not 'schema'.

Also modify AG and AM test to use this new property to only support
parsing, since these tests are not worth the effort or can't possibly be
made to support unparsing.

DFDL-1564