Do not abort when dumping utf-16/32 data
We currently abort when outputing utf-16/32 data because the byte order
mark is required the to know how to properly output the data. Except the
DFDL spec says that if no byte order mark exists, then to use the big
endian encoding. Daffodil does not support byte order marks, so it just
always uses the big endian encoding when the encoding is specified as
utf-16.

So, when dumping utf-16 data, it is consistent with the rest of Daffodil
to use the normal java utf-16 encoder, which will default to utf-16be.
Note that this will eventually need to change when we do add support for
BOM so that the encoder takes that into account.

DFDL-1402