When compiling DFDL schemas in the TDML Runner, the Processor cannot be serialized/deserialized if a test case has validation="on". Otherwise, a usage error will be thrown since full validation cannot work with saved/reloaded parsers.
It is possible that two test cases with different validation modes,
could reference the same DFDL schema, and thus the same Processor due to
the SchemaCache. If the Processor that was put into the cache first was
one that was serialized, test cases that share the schema source but
with validation="on" will now fail when they use the cached serialized
Processor.

This changes the SchemaCache key to be a tuple of the URISchemaSource
and whether or not the resulting Processor was serialized. This allows
test cases to get a serialized or non-serialized Processor based on the
validation mode of the test.

DFDL-1576