Fix incorrect calculation in content/value lengths
In the InfosetImpl, when the maybe{Start,End}DataOutputStream isDefined,
the maybe{Start,End}Pos0bInBits is a position relative to the beginning
of the DOS. When the absolute position of the data output stream
isDefined, we are able to calculate the absolute position of that
captured relative position. However, this calculation was incorrect
because it didn't take into account the relative position of the data
output stream. For example, assume the absolute position of the data
stream is 200 and the relative bit position is 50. That means this data
output stream started at 150, and so the captured relative position must
be added to 150, rather than subtracted from 200 (which is what this did
previously).