Calculate relative value/contentLengths when all DOS inbetween start/end are finished
When calculating value/contentLength, even if both captured start and
end bit positions are relative and in different data output streams, we
can still calculate the length if the start and end DOS's and all DOS's
in between are all finished. Add that ability, allowing previously
blocked suspensions to finish.

This allows the full ICMP test to round trip.

DFDL-1561