Add the ablity to mark a suspension as readOnly
If we know a suspension is read only and will never write to its data
output stream, then we can immediately mark the data output stream as
finished. This allows other suspensions waiting for the data output
stream to not be blocked on this one and helps to break some
circularies. Right now, the only suspension taking advantage of this is
the TargetLengthOperation.