Increase the byte buffer limit when filling a byte buffer that ends unaligned
In some cases, such as when a hexBinary field has a specified length and
the field is unaligned, the byte limit doesn't allow us to get all the
necessary data. This is because the limit of the data is set to (bitLimit >>
3), and maybeBitLimitOffset is set to bitLimit % 8. In order to get
those offset bytes in fillByteBufferUnaligned, we need to increase the
byte limit by one, get the data, and then reset the byte limit back to
where it was.