Reset the saved DPath current node in case something throws
When evaluating DPath arrays indicies, we save the current node before
running the expression that determines the index. That subexpression
resets the state, so the current node is null upon returning. If one of
the following withRetryIfBlocking calls throws (e.g.
InfosetArrayIndexOutOfBounds), then the current node remains null.
Future functions, such as fn:exists, will catch the thrown exception and
assert that the current node is not null. So we must reset the current
node to the saved node before calling the functions that could throw.

DFDL-1621