Fix USASCII 7Bit encoder to handle replacing unmappable characters
When trying to encode an unmappable character (i.e. a character that
needs 8 bits to represent), we returned a CoderResult that signified
this. However, the CharsetEncoder that receives this error assumes the
replacement character is made up of full bytes. With this encoding, that
is not the case. So modify the 7Bit encoder so that if
encodingErrorPolicy="replace", then it replaces the unmappable character
itself. The CharsetEncoder never knows there is an unmappable character,
which works out fine. Furthermore, if encodingErrorPolicy="error" then
we do report the unmappable error, but we must also back track the input
CharBuffer to before the error occurred, since the CharsetEncoder may
attempt to skip over the invalid character.

Also add some exception catching so that we throw UnparseErrors when a
putString failes due to an encoding error.

DFDL-1620