Throw a more user friendly error for unsupported lengthUnits for value/contentLength functions
Instead of throwing ???, throw an IllegalArgumentException specifying that
lengthUnits='characters' for dfdl:valueLength and dfdl:contentLength is
not yet supported, which results in a more friendly and consistent ParseError.

Modify tests to expect the not yet implemented error. When we eventually
add the support, those tests will fail and we'll remember to update
them.

DFDL-1626