Fix when CaptureValueLengthParsers are needed
CaptureValueLengthParsers are not needed on complex elements with
specified lengths. Instead, the specified length parsers handle
capturing the value length. However, the code that determines if
CaptureValueLengthParsres are needed or not treated complex elements as
lengthKind="delimited" as specified length. However, this is not the
case. Specified length parsrers are not inserted for complex types with
lengthKind="delimited". This means that for complex elements with
lengthKind="delimited", the value length would never be captured. This
fixes that so that if a complex element has lengthKind="delimited", we
will insert a CaptureValueLengthParser if necessary.

DFDL-1627