Fix fn:nilled to not require a value
The FNNilled class extended FNOneArg, which always attempts to retrieve
the value of the argument. This causes an exception since nilled
elements do not have a value. So instead, extend RecipeOpWithSubRecipe
and modify the run() method to try to get the node and check if it is
nilled. This does not attempt to check for existence and return an empty
sequence as the spec describes, since that use case isn't really
practical for us. Accepting an empty sequence as an argument or
returning an empty sequence doesn't really make sense. If it is possible
for an element to not exist, one should first check fn:exists() before
calling fn:nilled().

DFDL-1233
DFDL-1076