Allow controlling whether values must be set when writing the infoset
Before this change, when writing the infoset, we would always skip
writing simple elements values if the value was not set yet. This was
necessary because the debugger might want to write the infoset before a
value was set, and we would not want it to abort in that case. However,
this meant that if there was a bug in Daffodil and we did not properly
set a value, it would be possible that we would would just print nothing
for the element. Ideally, we would hit the code path that checks for set
values and throw an exception if a value was not set.

Furthermore, the code did not check isNilledSet for nillable elements,
which caused an abort in the debugger when writing nillable elements
that had not had a value set or had not had nil set yet.

To fix these issues, this adds a parameter to toWriter()
(allowUnsetValues) to control whether we should print nothing for unset
values or if we should always attempt to retrieve them and abort if they
have not been set. This configuration option is used when accessing both
simple element values or isNilled.

Note that this should only be set to true from the debugger, as it is a
useful way to catch errors in other cases. The debugger is also modified
so that when writing the result of an eval call it will obey the
debugger removeHidden setting and allow unset values, which is the same
behavior as when writing the infoset.

DFDL-1645