Use ComparisionExpression for comparing eq and ne
The logic in EqualityComparision was not correct when it came to
comparing numeric values. This is because the targetTypeForSubexpression
was NodeInfo.AnyAtomic, so the conversionOps did not think it needed to
do any conversions. However, if the operands are numeric, a conversion
is necessary to convert to like numeric types. This was not happening so
something like (xs:unsignedLong(0) eq xs:int(0)) would return false.

The ComparisonExpression already has the neccesary logic to figure out
the converged numeric types and set targetTypeForSubexpression
appropriately so that the type conversion works. Furthermore, the
special logic that used to exist in EqualityComparison no longer exists.
So remove the special casing of equalty comparissions and put its logic
in ComparisonExpression.

DFDL-1644