Allow fn:count() to unblock when unparsing when the element does not exists
In fn:count, it uses arrayFinalLength to block on an array until the
array becomes final. However, if the array has a length of zero, then it
will never exist and we will never unblock because we can't ask the
array if it is final.

So, modify fn:count so it behaves basically like fn:exists, except if it
does exist, return the finalArrayLength. If we determine it does not
exist, return 0. This way the exists logic is used to determine if an
array exists and if it does, use the finalArrayLength to determine when
the array becomes final.

DFDL-1648