Add check for removing some alignment parsers
The isKnownToBePrecededByAllByteLengths variable is supposed to
determine when alignment fill parsers/unparsers are necessary by
statically determining when elements are always going to be byte
aligned. However, this is currently does not remove many parsers, and is
actually quite difficult to fix due to the difficulty in determine what
elements could precede other elements.

Instead of fixing that, this patch adds a new variable,
isAllKnownToBeByteAlignedAndByteLength, which determines if all elements
are byte aligned, have a byte length, and have a non-byte-size encoding.
If all elements have these properties, then we know everything is going
to be byte aligned and alignment fill parsers/unparsers are not
necessary. Note that this is not going to optimize formats that have
mixed alignments, bit lengths, or non-byte-size encodings, but it should
help in the common case for files that are all byte aligned with byte
lengths.

In addition to removing parsers/unparsers and improving performance,
this also makes it much easier to avoid deadlocks during unparsing,
since alignment unparsers block and need all prior data to be written
before unblocking.

DFDL-1650