Fix abort when with repeated occurrences, a discriminator is true, and a subparser fails
In this specific case, when we have repeated corrances (either N or
unbounded), a discriminator gets set to true, and a subparser fails, we
attempt to reset back to the startState mark, essentially ignoring all
work done during the repeats. Immediately after resetting the startState
we also discard the priorState mark. However, the priorState mark is
taken after the startState, so resetting to startState implicitly
discards priorState. By then attempting to then discard priorState, it
causes an abort since it cannot find priorState in the mark list.

This patch removes the priorState discard, since that is implicitly done
by the startState reset.

DFDL-1653