Enable capturing dfdl:value/contextLength of element references
For performance reasons, we only insert CaptureValue/ContentLength
unparsers when we determine that an element is used in
dfdl:value/contentLength function. We do this by building up a list of
all dpathCompileInfo's that are used in such functions. Then we check to
see if the dpathCompileInfo of each element is inside that list. If it
is, then we insert the appropriate parsers/unparsers.

However, if there was an elementRef, we would actually create two
dpathCompileInfo's, one for the element and one for the elementRef. But
only the dpathCompileInfo's for elements (not elementRefs) are stored in
the list. This meant that dpathCompileInfo's for elementRefs were not in
the list, and so we would not capture their lengths.

This changes this so that asking for the dpathCompileInfo of an
elementRef just returns the dpathCompileInfo of the referenced element,
rather than creating a new one. This results in correctly determining
that element refs are used inside value/contentLength functions, and the
appropriate capture parsers are inserted.

DFDL-1657